JMS and DBSeeder fix

+++ web-eshop/src/main/java/cz/kodytek/shop/presentation/controllers/InvoiceDetailController.java
@@ -3,6 +3,7 @@ package cz.kodytek.shop.presentation.controllers;
 import cz.kodytek.shop.data.entities.interfaces.invoice.IInvoice;
 import cz.kodytek.shop.data.entities.invoice.InvoiceGood;
 import cz.kodytek.shop.domain.services.interfaces.invoices.IInvoicesService;
+import cz.kodytek.shop.jms.JMSService;
 import cz.kodytek.shop.presentation.session.models.FlashMessage;
 import cz.kodytek.shop.presentation.session.models.FlashMessageType;
 import cz.kodytek.shop.presentation.session.services.interfaces.messages.IFlashMessagesService;
@@ -26,12 +27,17 @@ public class InvoiceDetailController {
     @Inject
     private IFlashMessagesService flashMessagesService;
 
+    @Inject
+    private JMSService jmsService;
+
     private IInvoice invoice;
 
     public IInvoice getInvoice() {
         if (requestUtils.hasParam("number")) {
-            if (invoice == null)
+            if (invoice == null) {
+                jmsService.sendMessage(requestUtils.getIp() + " attempting to load invoice, " + requestUtils.getParam("number") + ".");
                 invoice = invoicesService.getForInvoiceNumber(requestUtils.getParam("number"));
+            }
             if (invoice != null) {
                 return invoice;
             } else
@@ -43,19 +49,22 @@ public class InvoiceDetailController {
     }
 
     public void delete() {
-        invoicesService.deleteForInvoice(requestUtils.getParam("number"));
+        if (requestUtils.hasParam("number")) {
+            jmsService.sendMessage(requestUtils.getIp() + " deleting invoice, " + requestUtils.getParam("number") + ".");
+            invoicesService.deleteForInvoice(requestUtils.getParam("number"));
+        }
         requestUtils.redirect("/pages/index.xhtml", new FlashMessage("Invoice successfully canceled.", FlashMessageType.success));
     }
 
     public int getUnitCount() {
-        if(getInvoice() == null)
+        if (getInvoice() == null)
             return 0;
 
         return getInvoice().getGoods().stream().mapToInt(InvoiceGood::getAmount).sum();
     }
 
     public Money getPriceCount() {
-        if(getInvoice() == null)
+        if (getInvoice() == null)
             return Money.of(0, "CZK");
 
         return getInvoice().getGoods().stream().map(i -> i.getCost().multiply(i.getAmount())).reduce(Money::add).get()
